feat: add MagicOnion OpenTelemetry Server Sample
