+++ samples/ChatApp/ChatApp.Server.TelemetryNew/appsettings.json
@@ -1,18 +1,23 @@
 {
-  "MagicOnion": {
-    "ServerPorts": [
-      {
-        "Host": "0.0.0.0",
-        "Port": 12345,
-        "UseInsecureConnection": true
-      }
-    ]
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Trace",
-      "System": "Information",
-      "Microsoft": "Information"
+    "MagicOnion": {
+        "ServerPorts": [
+            {
+                "Host": "0.0.0.0",
+                "Port": 12345,
+                "UseInsecureConnection": true
+            }
+        ],
+        "OpenTelemetry": {
+            "ServiceName": "ChatApp.Server.Telemery",
+            "MetricsExporterEndpoint": "http://127.0.0.1:9181/metrics/",
+            "TracerExporterEndpoint": "http://127.0.0.1:9411/api/v2/spans"
+        }
+    },
+    "Logging": {
+        "LogLevel": {
+            "Default": "Trace",
+            "System": "Information",
+            "Microsoft": "Information"
+        }
     }
-  }
 }
