+++ samples/ChatApp/ChatApp.Server.TelemetrySample/ChatService.cs
@@ -1,0 +1,22 @@
+﻿using ChatApp.Shared.Services;
+using MagicOnion;
+using MagicOnion.Server;
+using MessagePack;
+
+namespace ChatApp.Server
+{
+    public class ChatService : ServiceBase<IChatService>, IChatService
+    {
+        public UnaryResult<Nil> GenerateException(string message)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public UnaryResult<Nil> SendReportAsync(string message)
+        {
+            Logger.Debug($"{message}");
+
+            return UnaryResult(Nil.Default);
+        }
+    }
+}
