+++ samples/ChatApp/ChatApp.Server.TelemetrySample/appsettings.Development.json
@@ -1,0 +1,14 @@
+{
+  "MagicOnion": {
+    "Service": {
+      "IsReturnExceptionStackTraceInErrorDetail": true
+    },
+    "ServerPorts": [
+      {
+        "Host": "localhost",
+        "Port": 12345,
+        "UseInsecureConnection": true
+      }
+    ]
+  }
+}
