+++ samples/ChatApp/ChatApp.Server.TelemetrySample/appsettings.json
@@ -1,0 +1,18 @@
+{
+  "MagicOnion": {
+    "ServerPorts": [
+        {
+            "Host": "0.0.0.0",
+            "Port": 12345,
+            "UseInsecureConnection": true
+        }
+    ]
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Trace",
+      "System": "Information",
+      "Microsoft": "Information"
+    }
+  }
+}
